+++ Editor/Analysis/ActionsRegistrationGenerator~/ActionsGenerator.cs
@@ -24,11 +24,56 @@ public class ActionRegistrationSourceGenerator : ISourceGenerator
     public void Execute(GeneratorExecutionContext context)
     {
         var output = GetOutput(context);
+        output.WriteLine(DateTime.Now);
+
+        // we need to know if the settings are configured to not perform codegen to link attributed methods
+        // this is kinda annoying because the path root of the project settings and the root path of this process are *very* different
+        // so what we do is we use the included Compilation Assembly additional file that Unity gives us.
+        // This file if opened has the path of the Unity project, which we can then use to get the settings
+        // if any stage of this fails then we bail out and assume that codegen is desired
+        if (context.AdditionalFiles.Any())
+        {
+            var relevants = context.AdditionalFiles.Where(i => i.Path.Contains($"{context.Compilation.AssemblyName}.AdditionalFile.txt"));
+            if (relevants.Any())
+            {
+                var arsgacsaf = relevants.First();
+                if (File.Exists(arsgacsaf.Path))
+                {
+                    try
+                    {
+                        var projectPath = File.ReadAllText(arsgacsaf.Path);
+                        var fullPath = Path.Combine(projectPath, Yarn.Unity.Editor.YarnSpinnerProjectSettings.YarnSpinnerProjectSettingsPath);
+                        output.WriteLine($"Attempting to read settings file at {fullPath}");
+
+                        if (!Yarn.Unity.Editor.YarnSpinnerProjectSettings.GetOrCreateSettings(projectPath, output).automaticallyLinkAttributedYarnCommandsAndFunctions)
+                        {
+                            output.WriteLine("Skipping codegen due to settings.");
+                            output.Dispose();
+                            return;
+                        }
+                    }
+                    catch (Exception e)
+                    {
+                        output.WriteLine($"Unable to determine Yarn settings, settings values will be ignored and codegen will occur: {e.Message}");
+                    }
+                }
+                else
+                {
+                    output.WriteLine($"The project settings path metadata file does not exist at: {arsgacsaf.Path}. Settings values will be ignored and codegen will occur");
+                }
+            }
+            else
+            {
+                output.WriteLine("Unable to determine Yarn settings path, no file containing the project path metadata was included. Settings values will be ignored and codegen will occur.");
+            }
+        }
+        else
+        {
+            output.WriteLine("Unable to determine Yarn settings path as no additional files were included. Settings values will be ignored and codegen will occur.");
+        }
         
         try
         {
-            output.WriteLine(DateTime.Now);
-
             output.WriteLine("Source code generation for assembly " + context.Compilation.AssemblyName);
 
             if (context.AdditionalFiles.Any()) {
@@ -96,14 +141,17 @@ public class ActionRegistrationSourceGenerator : ISourceGenerator
             var actions = new List<YarnAction>();
             foreach (var tree in compilation.SyntaxTrees)
             {
-                actions.AddRange(Analyser.GetActions(compilation, tree).Where(a => a.DeclarationType == DeclarationType.Attribute));
+                actions.AddRange(Analyser.GetActions(compilation, tree, output).Where(a => a.DeclarationType == DeclarationType.Attribute));
             }
 
-            if (actions.Any() == false) {
+            if (actions.Any() == false)
+            {
                 output.WriteLine($"Didn't find any Yarn Actions in {context.Compilation.AssemblyName}. Not generating any source code for it.");
                 return;
             }
 
+            HashSet<string> removals = new HashSet<string>();
+            // validating and logging all the actions
             foreach (var action in actions)
             {
                 if (action == null)
@@ -111,10 +159,60 @@ public class ActionRegistrationSourceGenerator : ISourceGenerator
                     output.WriteLine($"Action is null??");
                     continue;
                 }
+
+                // if an action isn't public we will log a warning
+                // and then later we will also skip it
+                if (action.MethodSymbol.DeclaredAccessibility != Accessibility.Public)
+                {
+                    var descriptor = new DiagnosticDescriptor(
+                        "YS1001",
+                        $"Yarn {action.Type} methods must be public",
+                        "YarnCommand and YarnFunction methods must be public. \"{0}\" is {1}.",
+                        "Yarn Spinner",
+                        DiagnosticSeverity.Warning,
+                        true,
+                        "[YarnCommand] and [YarnFunction] attributed methods must be public so that the codegen can reference them.",
+                        "https://docs.yarnspinner.dev/using-yarnspinner-with-unity/creating-commands-functions");
+                    context.ReportDiagnostic(Microsoft.CodeAnalysis.Diagnostic.Create(
+                        descriptor,
+                        action.Declaration?.GetLocation(),
+                        action.MethodIdentifierName, action.MethodSymbol.DeclaredAccessibility
+                    ));
+                    output.WriteLine($"Action {action.Name} will be skipped due to it's declared accessibility {action.MethodSymbol.DeclaredAccessibility}");
+                    removals.Add(action.Name);
+                    continue;
+                }
+
+                // this is not a full validation of the naming rules of commands
+                // but is good enough to catch the most common situations, whitespace and periods
+                if (action.Name.Contains(".") || action.Name.Any(x => Char.IsWhiteSpace(x)))
+                {
+                    var descriptor = new DiagnosticDescriptor(
+                        "YS1002",
+                        $"Yarn {action.Type} methods must have a valid name",
+                        "YarnCommand and YarnFunction methods follow existing ID rules for Yarn. \"{0}\" is invalid.",
+                        "Yarn Spinner",
+                        DiagnosticSeverity.Warning,
+                        true,
+                        "[YarnCommand] and [YarnFunction] attributed methods must follow Yarn ID rules so that Yarn scripts can reference them.",
+                        "https://docs.yarnspinner.dev/using-yarnspinner-with-unity/creating-commands-functions");
+                    context.ReportDiagnostic(Microsoft.CodeAnalysis.Diagnostic.Create(
+                        descriptor,
+                        action.Declaration?.GetLocation(),
+                        action.Name
+                    ));
+                    output.WriteLine($"Action {action.MethodIdentifierName} will be skipped due to it's name {action.Name}");
+                    removals.Add(action.Name);
+                    continue;
+                }
+
                 output.WriteLine($"Action {action.Name}: {action.SourceFileName}:{action.Declaration?.GetLocation()?.GetLineSpan().StartLinePosition.Line} ({action.Type})");
             }
 
-            output.Write($"Generating source code... ");
+            // removing any actions that failed validation
+            actions.Where(x => !removals.Contains(x.Name)).ToList();
+
+            output.Write($"Generating source code...");
 
             var source = Analyser.GenerateRegistrationFileSource(actions);
 
@@ -342,18 +440,40 @@ public class ActionRegistrationSourceGenerator : ISourceGenerator
         context.RegisterForSyntaxNotifications(() => new ClassDeclarationSyntaxReceiver());
     }
 
-    public ILogger GetOutput(GeneratorExecutionContext context)
+    static string TemporaryPath()
+    {
+        var tempPath = System.IO.Path.Combine(System.IO.Path.GetTempPath(), "dev.yarnspinner.logs");
+
+        // we need to make the logs folder, but this can potentially fail
+        // if it does fail then we will just chuck the logs inside the tmp folder
+        try
+        {
+            if (!Directory.Exists(tempPath))
+            {
+                Directory.CreateDirectory(tempPath);
+            }
+        }
+        catch
+        {
+            tempPath = System.IO.Path.GetTempPath();
+        }
+        return tempPath;
+    }
+
+    public Yarn.Unity.ILogger GetOutput(GeneratorExecutionContext context)
     {
         if (GetShouldLogToFile(context))
         {
-            var tempPath = System.IO.Path.GetTempPath();
-            var path = System.IO.Path.Combine(tempPath, $"{nameof(ActionRegistrationSourceGenerator)}-{context.Compilation.AssemblyName}.txt");
+            var tempPath = ActionRegistrationSourceGenerator.TemporaryPath();
 
+            var path = System.IO.Path.Combine(tempPath, $"{nameof(ActionRegistrationSourceGenerator)}-{context.Compilation.AssemblyName}.txt");
             var outFile = System.IO.File.Open(path, System.IO.FileMode.Create);
 
-            return new FileLogger(new System.IO.StreamWriter(outFile));
-        } else {
-            return new NullLogger();
+            return new Yarn.Unity.FileLogger(new System.IO.StreamWriter(outFile));
+        }
+        else
+        {
+            return new Yarn.Unity.NullLogger();
         }
     }
 
@@ -364,8 +484,9 @@ public class ActionRegistrationSourceGenerator : ISourceGenerator
 
     public void DumpGeneratedFile(GeneratorExecutionContext context, string text)
     {
-        if (GetShouldLogToFile(context)) {
-            var tempPath = System.IO.Path.GetTempPath();
+        if (GetShouldLogToFile(context))
+        {
+            var tempPath = ActionRegistrationSourceGenerator.TemporaryPath();
             var path = System.IO.Path.Combine(tempPath, $"{nameof(ActionRegistrationSourceGenerator)}-{context.Compilation.AssemblyName}.cs");
             System.IO.File.WriteAllText(path, text);
         }
@@ -383,54 +504,5 @@ internal class ClassDeclarationSyntaxReceiver : ISyntaxReceiver
         {
             Classes.Add(cds);
         }
-    }
-}
-
-public interface ILogger : IDisposable
-{
-    void Write(object obj);
-    void WriteLine(object obj);
-}
-
-public class FileLogger : ILogger
-{
-    System.IO.TextWriter writer;
-
-    public FileLogger(TextWriter writer)
-    {
-        this.writer = writer;
-    }
-
-    public void Dispose()
-    {
-        writer.Dispose();
-    }
-
-    public void Write(object text)
-    {
-        writer.Write(text);
-    }
-
-    public void WriteLine(object text)
-    {
-        writer.WriteLine(text);
-    }
-}
-
-public class NullLogger : ILogger
-{
-    public void Dispose()
-    {
-
-    }
-
-    public void Write(object text)
-    {
-
-    }
-
-    public void WriteLine(object text)
-    {
-
     }
 }