+++ Editor/Analysis/Analyser.cs
@@ -409,8 +409,14 @@ namespace Yarn.Unity.ActionAnalyser
             }
         }
 
-        public static IEnumerable<Action> GetActions(CSharpCompilation compilation, Microsoft.CodeAnalysis.SyntaxTree tree)
+        public static IEnumerable<Action> GetActions(CSharpCompilation compilation, Microsoft.CodeAnalysis.SyntaxTree tree, Yarn.Unity.ILogger yLogger = null)
         {
+            var logger = yLogger;
+            if (logger == null)
+            {
+                logger = new NullLogger();
+            }
+
             var root = tree.GetCompilationUnitRoot();
 
             SemanticModel model = null;
@@ -419,7 +425,7 @@ namespace Yarn.Unity.ActionAnalyser
                 model = compilation.GetSemanticModel(tree, true);
             }
 
-            return GetAttributeActions(root, model).Concat(GetRuntimeDefinedActions(root, model));
+            return GetAttributeActions(root, model, logger).Concat(GetRuntimeDefinedActions(root, model));
         }
 
         private static IEnumerable<Action> GetRuntimeDefinedActions(CompilationUnitSyntax root, SemanticModel model)
@@ -526,9 +532,8 @@ namespace Yarn.Unity.ActionAnalyser
             }
         }
 
-        private static IEnumerable<Action> GetAttributeActions(CompilationUnitSyntax root, SemanticModel model)
+        private static IEnumerable<Action> GetAttributeActions(CompilationUnitSyntax root, SemanticModel model, Yarn.Unity.ILogger logger)
         {
-
             var methodInfos = root
                 .DescendantNodes()
                 .OfType<MethodDeclarationSyntax>()
@@ -539,8 +544,7 @@ namespace Yarn.Unity.ActionAnalyser
                 {
                     return (MethodDeclaration: decl, Symbol: model.GetDeclaredSymbol(decl));
                 })
-                .Where(pair => pair.Symbol != null)
-                .Where(pair => pair.Symbol.DeclaredAccessibility == Accessibility.Public);
+                .Where(pair => pair.Symbol != null);
 
             var actionMethods = methodsAndSymbols
                 .Select(pair =>
@@ -555,21 +559,29 @@ namespace Yarn.Unity.ActionAnalyser
             {
                 var attr = methodInfo.ActionAttribute;
 
-                string actionName;
+                // working on an assumption that most people just use the method name
+                string actionName = methodInfo.MethodDeclaration.Identifier.ToString();
 
-                // If the attribute has an argument list, and the first argument
-                // is a string, then use that string's value as the action name.
-                if (attr.ArgumentList != null
-                    && attr.ArgumentList.Arguments.First().Expression is LiteralExpressionSyntax commandName
-                    && commandName.Kind() == SyntaxKind.StringLiteralExpression
-                    && commandName.Token.Value is string name)
+                // handling the situation where they have provided arguments
+                // if we have an argument list
+                if (attr.ArgumentList != null)
                 {
-                    actionName = name;
-                }
-                else
-                {
-                    // Otherwise, use the method's name.
-                    actionName = methodInfo.MethodDeclaration.Identifier.ToString();
+                    // we resolve the value of first item in that list
+                    // and if it's a string we use that as the action name
+                    var constantValue = model.GetConstantValue(attr.ArgumentList.Arguments.First().Expression);
+                    if (constantValue.HasValue)
+                    {
+                        if (constantValue.Value is string)
+                        {
+                            logger.WriteLine($"resolved constant expression value for the action name: {constantValue.Value.ToString()}");
+                            actionName = constantValue.Value as string;
+                        }
+                        else
+                        {
+                            // Otherwise just logging the incorrect type and moving on with our life
+                            logger.WriteLine($"resolved constant expression value for the action name, but it is not a string, skipping: {constantValue.Value.ToString()}");
+                        }
+                    }
                 }
 
                 var position = methodInfo.MethodDeclaration.GetLocation();
@@ -584,6 +596,7 @@ namespace Yarn.Unity.ActionAnalyser
                     Name = actionName,
                     Type = methodInfo.ActionType,
                     MethodName = $"{containerName}.{methodInfo.MethodDeclaration.Identifier}",
+                    MethodIdentifierName = methodInfo.MethodDeclaration.Identifier.ToString(),
                     MethodSymbol = methodInfo.Symbol,
                     IsStatic = methodInfo.Symbol.IsStatic,
                     Declaration = methodInfo.MethodDeclaration,
@@ -628,8 +641,6 @@ namespace Yarn.Unity.ActionAnalyser
             // default value; other parts of the action detection process will throw
             // errors.
             return default;
-
-
         }
 
         private static IEnumerable<Parameter> GetParameters(IMethodSymbol symbol)